+++ src/main/java/step3/Application.java
@@ -1,5 +1,8 @@
 package step3;
 
+import step3.domain.Car;
+import step3.domain.CarFactory;
+import step3.service.RacingManager;
 import step3.view.InputView;
 import step3.view.OutputView;
