+++ src/main/java/step3/service/MovementConditionProvider.java
@@ -1,4 +1,4 @@
-package step3;
+package step3.service;
 
 import java.util.Random;
