+++ src/main/java/step3/service/MovementDecisionMaker.java
@@ -1,4 +1,4 @@
-package step3;
+package step3.service;
 
 public final class MovementDecisionMaker {
