+++ src/test/java/step3/domain/CarFactoryTest.java
@@ -1,7 +1,9 @@
-package step3;
+package step3.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import step3.domain.Car;
+import step3.domain.CarFactory;
 
 import java.util.List;
