+++ src/test/java/step3/domain/CarTest.java
@@ -1,9 +1,10 @@
-package step3;
+package step3.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
+import step3.domain.Car;
 
 import java.util.stream.Stream;
