+++ src/test/java/step3/service/MovementDecisionMakerTest.java
@@ -1,10 +1,11 @@
-package step3;
+package step3.service;
 
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
+import step3.service.MovementDecisionMaker;
 
 import java.util.stream.Stream;
