+++ src/test/java/step3/service/RacingManagerTest.java
@@ -1,10 +1,13 @@
-package step3;
+package step3.service;
 
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.mockito.MockedStatic;
+import step3.domain.Car;
+import step3.service.MovementConditionProvider;
+import step3.service.RacingManager;
 
 import java.util.ArrayList;
 import java.util.List;
